rename files to match usage
